Continued adding JSON to scripts - finished

+++ src/main/java/Controllers/UserController.java
@@ -2,6 +2,7 @@ package Controllers;
 
 import Server.Main;
 import com.sun.jersey.multipart.FormDataParam;
+import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 
 import javax.ws.rs.*;
@@ -12,7 +13,7 @@ import java.sql.ResultSet;
 public class UserController {
 
     @POST
-    @Path("new")
+    @Path("insertUser")
     @Consumes(MediaType.MULTIPART_FORM_DATA)
     @Produces(MediaType.APPLICATION_JSON)
     public String insertUser(
@@ -44,7 +45,7 @@ public class UserController {
     @Produces(MediaType.APPLICATION_JSON)
     public String selectUser(@PathParam("id") Integer id) throws Exception {
         if (id == null) {
-            throw new Exception("Thing's 'id' is missing in the HTTP request's URL.");
+            throw new Exception("User's 'id' is missing in the HTTP request's URL.");
         }
         System.out.println("thing/get/" + id);
         JSONObject item = new JSONObject();
@@ -67,20 +68,26 @@ public class UserController {
 
         }
     }
-
-    public static void selectAllUsers() {
+    @GET
+    @Path("selectAllUsers")
+    @Produces(MediaType.APPLICATION_JSON)
+    public String selectAllUsers() {
+        System.out.println("users/selectAllUsers");
+        JSONArray list = new JSONArray();
         try {
             PreparedStatement ps = Main.db.prepareStatement("SELECT * FROM Users ");
             ResultSet results = ps.executeQuery();
             while (results.next()) {
-                int UserID = results.getInt(1);
-                String Name = results.getString(2);
-                int BestScore = results.getInt(3);
-                String Password = results.getString(4);
-                System.out.println(UserID + " " + Name + " " + BestScore + " " + Password);
-            }
+                JSONObject item = new JSONObject();
+                item.put("UserID",results.getString(1));
+                item.put("Name", results.getString(2));
+                item.put("BestScore", results.getInt(3));
+                item.put("Password", results.getString(4));
+              }
+            return list.toString();
         } catch (Exception exception) {
             System.out.println("Database error: " + exception.getMessage());
+            return "{\"error\": \"Unable to list items, please see server console for more info.\"}";
         }
     }
 
@@ -102,13 +109,25 @@ public class UserController {
         }
     }
 
-    public static void deleteDatabase(int UserID) {
+    @POST
+    @Path("deleteUser")
+    @Consumes(MediaType.MULTIPART_FORM_DATA)
+    @Produces(MediaType.APPLICATION_JSON)
+    public String deleteUser(@FormDataParam("UserId") Integer UserID) {
         try {
+            if (UserID == null) {
+                throw new Exception("One or more form data parameters are missing in the HTTP request.");
+            }
+            System.out.println("Users/deleteUser id=" + UserID);
+
             PreparedStatement ps = Main.db.prepareStatement("DELETE FROM Users WHERE UserID = ?)");
             ps.setInt(1, UserID);
             ps.executeUpdate();
+            return "{\"status\": \"OK\"}";
         } catch (Exception exception) {
             System.out.println("Database error: " + exception.getMessage());
+            return "{\"error\": \"Unable to delete item, please see server console for more info.\"}";
+
         }
     }
 }
